split failpoints and integration tests
